Create Lab_01
Consider a household trading agent that monitors the price of some commodity, such as toilet paper, by checking for deals online, as well as how much the household has in stock. It must decide whether to order more and how much to order. Assume the percepts are the price and the amount in stock. The command is the number of units the agent decides to order (which is zero if the agent does not order any). A percept trace specifies for each time point (e.g., each day) the price at that time and the amount in stock at that time.
1. If the last_price is less than 70% of the average (ave) and the instock quantity is less than 30, then set tobuy to 15.
2. Otherwise, if the instock quantity is less than 10, set tobuy to 5. If none of the above conditions apply, set tobuy to 0.